requirements.txt
# Core
fastapi
uvicorn
streamlit
pydantic
python-dotenv

# LLM and Embeddings
langchain
langchain-community
langchain-huggingface
sentence-transformers
torch
google-generativeai

# Vector Store
chromadb
duckdb

# Crawler and Parsing
httpx
beautifulsoup4
html2text

# Neo4j for KG
neo4j

# Optional (logging, retry, etc.)
tqdm
